:truck: move legend tests into new test file
